pyproject.toml

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "log10-io"
version = "0.4.5"
authors = ["log10 team"]
license = "MIT"
description = "Unified LLM data management"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
packages = [
    { include = "log10" },
]


[tool.poetry.group.dev.dependencies]
build = "^0.10.0"
ruff = "^0.1.6"
black = "^23.11.0"
mypy = "^1.7.0"

[project.urls]
"Homepage" = "https://github.com/log10-io/log10"
"Bug Tracker" = "https://github.com/log10-io/log10/issues"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langchain = "^0.0.303"
openai = "^0.28.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
google-cloud-bigquery = "^3.11.4"
pexpect = "^4.8.0"
faiss-cpu = "^1.7.4"
google-search-results = "^2.4.2"
tiktoken = "^0.4.0"
wikipedia = "^1.4.0"
faker = "^19.2.0"
backoff = "^2.2.1"
anthropic = "^0.3.11"
mosaicml-cli = "^0.5.30"
together = "^0.2.7"
types-requests = "^2.31.0.10"

[tool.isort]
profile = "black"
length_sort = true

[tool.ruff]
line-length = 120
target-version = "py39"
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
unfixable = [
  # disable auto fix for print statements
  "T201",
  "T203",
]
ignore-init-module-imports = true
[tool.ruff.per-file-ignores]
"log10/agents/camel.py" = ["E501"]
